chore: module list layout consistency
