+++ web/components/modules/index.ts
@@ -12,6 +12,7 @@ export * from "./module-card-item";
 export * from "./module-list-item";
 export * from "./module-peek-overview";
 export * from "./quick-actions";
+export * from "./module-list-item-action";
 
 // archived modules
 export * from "./archived-modules";
