+++ web/components/modules/modules-list-view.tsx
@@ -2,6 +2,7 @@ import { observer } from "mobx-react-lite";
 import Image from "next/image";
 import { useRouter } from "next/router";
 // components
+import { ListLayout } from "@/components/core/list";
 import { EmptyState } from "@/components/empty-state";
 import { ModuleCardItem, ModuleListItem, ModulePeekOverview, ModulesListGanttChartView } from "@/components/modules";
 import { CycleModuleBoardLayout, CycleModuleListLayout, GanttLayoutLoader } from "@/components/ui";
@@ -68,12 +69,17 @@ export const ModulesListView: React.FC = observer(() => {
     <>
       {displayFilters?.layout === "list" && (
         <div className="h-full overflow-y-auto">
-          <div className="flex h-full w-full justify-between">
-            <div className="flex h-full w-full flex-col overflow-y-auto vertical-scrollbar scrollbar-lg">
+          <div className="flex h-full w-full flex-col">
+            <div className="h-[50px] flex-shrink-0 w-full border-b border-custom-border-200 px-6 relative flex items-center gap-4 justify-between">
+              <div className="flex items-center">
+                <span className="block text-sm font-medium">Module name</span>
+              </div>
+            </div>
+            <ListLayout>
               {filteredModuleIds.map((moduleId) => (
                 <ModuleListItem key={moduleId} moduleId={moduleId} />
               ))}
-            </div>
+            </ListLayout>
             <ModulePeekOverview
               projectId={projectId?.toString() ?? ""}
               workspaceSlug={workspaceSlug?.toString() ?? ""}
